refactor: (tobe) 수량 Value Object common 패키지 이동
